requirements.txt
# Core dependencies
streamlit>=1.24.0
langchain>=0.0.267
openai>=0.27.8
pandas>=2.0.3
numpy>=1.24.3

# Web scraping
beautifulsoup4>=4.12.2
selenium>=4.10.0
webdriver-manager>=3.8.6
requests>=2.31.0
scrapy>=2.9.0
playwright>=1.36.0

# Data processing
openpyxl>=3.1.2
python-dotenv>=1.0.0
pytube>=15.0.0
pillow>=10.0.0
moviepy>=1.0.3
